requirements.txt
numpy
tensorflow
matplotlib
opencv-python
flask
pillow